requirements.txt
fastapi
uvicorn
twilio
langchain
langchain-google-genai
langchain-huggingface
langchain-community
langchain-qdrant
qdrant-client
sentence-transformers
pymupdf
beautifulsoup4
requests
python-dotenv
python-multipart
ragas
datasets
langchain-text-splitters
langchain-community
